removed ${yaml-cpp-pm_LIB} from EXTERNAL_LIBS for installation in case of a shared library build.
The installation will depend on the build files in case of a static library when using the contribute yaml library, because we don't install it.

Maybe we should fix that too? But I don't think this is worth the effort, because a static installation is probably pretty unusual and only used by experts who can easily recover from that problem.