Histogram<T>::computeStats() without sorting
Currently all values added to a histogram are sorted before being written to a csv file, but I need the values to be written in the order in which they were added.   

## Reason

All histogram value are written in a stream at `pointmatcher/Histogram.cpp#L80`. However, `computeStats()` is caled just before and do a `std::nth_element()` on the histogram, so it becomes sorted.

## Solution

I suggest calculating the median, upper and lower quarter on a copy of the histogram, so that it is not sorted.